Use new Snapshot class everywhere.
Since popup.js is an injected content script and can't use imports, I
couldn't figure out a nice way to reuse the same Snapshot class defined
in utils.js. The stopgap solution is duplicating the definition in
popup.js. This is cheap and simple, if somewhat clumsy. Maybe worth
revisiting soon.

Also, we're keeping the name as 'entries' in local storage to keep users' previous snapshots.